+++ packages/gov-ui-kit/src/modules/assets/copy/modulesCopy.ts
@@ -27,6 +27,8 @@ export const modulesCopy = {
     },
     proposalDataListItemStatus: {
         voted: "You've voted",
+        ago: 'ago',
+        left: 'left',
     },
     proposalDataListItemStructure: {
         by: 'By',